requirements.txt
requests
playwright
firecrawl-py
xhshow
itchat-uos
pycryptodome
zstandard
